# Change ShredIndex to use BitVector for underlying storage