--- a/ledger/Cargo.toml
+++ b/ledger/Cargo.toml
@@ -11,6 +11,7 @@ edition = "2018"
 
 [dependencies]
 bincode = "1.3.3"
+bv = { version = "0.11.1", features = ["serde"] }
 byteorder = "1.4.3"
 chrono = { version = "0.4.11", features = ["serde"] }
 chrono-humanize = "0.2.1"