--- a/ledger/src/blockstore.rs
+++ b/ledger/src/blockstore.rs
@@ -5864,10 +5864,10 @@ pub mod tests {
                     DataShredHeader::default(),
                     coding.clone(),
                 );
-                coding_shred.common_header.fec_set_index = std::u32::MAX - 1;
+                coding_shred.common_header.fec_set_index = MAX_DATA_SHREDS_PER_SLOT as u32 - 1;
                 coding_shred.coding_header.num_coding_shreds = 3;
-                coding_shred.common_header.index = std::u32::MAX - 1;
-                assert!(!Blockstore::should_insert_coding_shred(
+                coding_shred.common_header.index = MAX_DATA_SHREDS_PER_SLOT as u32 - 1;
+                assert!(Blockstore::should_insert_coding_shred(
                     &coding_shred,
                     &last_root
                 ));