#### Problem
The blockstore uses ShredIndex structures to track whether shreds have been received or not. The elements that we store are the indices (u64) themselves so we can use a BitVector instead of a BTreeSet.

Here are some results for running the benchmark (on my machine) that is included as part of this change; I ran with size of 500 and 5000 (these sizes would be the slot size in shreds). The reported execution times would be per slot.

**Current tip of master**
```
Results for 500
test bench_shred_index_is_present  ... bench:         743 ns/iter (+/- 62)
test bench_shred_index_set_present ... bench:      31,303 ns/iter (+/- 5,349)
Results for 5000
test bench_shred_index_is_present  ... bench:       7,282 ns/iter (+/- 612)
test bench_shred_index_set_present ... bench:     460,156 ns/iter (+/- 19,850)
```
**this branch**
```
Results for 500
test bench_shred_index_is_present  ... bench:         801 ns/iter (+/- 65)
test bench_shred_index_set_present ... bench:       2,860 ns/iter (+/- 184)
Results for 5000
test bench_shred_index_is_present  ... bench:       8,850 ns/iter (+/- 303)
test bench_shred_index_set_present ... bench:      28,112 ns/iter (+/- 4,920)
```
Fairly comparable on `is_present()`; 10x speedup for `set_present()` at slot size of 500 (and more pronounced at 5000), although it isn't a large chunk of time to start with. This should save us some space too though; with a a BTreeSet<u64> and assuming 500 shreds / slot, that would be 4k bytes / slot for each index, 8k total (data and coding). Hypothetically, should be some time savings in serialize/deserialize too.

### Upgrade / Downgrade
This is a breaking change since the ShredIndex gets serialized out of Rocks. Per some suggestions in this PR and conversation in Discord, I have implemented the following strategy:
- Introduce a second column family that uses BitVec, keeping the original BTreeSet column around
- Make a backport PR that stubs new column into older releases (this still creates a break point, but hopefully less painful)
   - Older validator binary would open DB with new column (RocksDB requires specifying all columns at open), but never use that column (except in purge maybe)

In the future once this change is reflected on a given network ...
- Strip out the original (BTreeSet) CF and only use the new (BitVec) one. Accomplish this with a feature switch (disabled state of feature writes both columns, enabled state of feature will only write new column)