@carllin 2021-07-14
> This would be a breaking change since ShredIndex gets serialized/deserialized into/out of Rocks; I'm unsure how we should go about rolling it out.

You could introduce a separate column family to hold this new index structure, and then we can deprecate and delete the old index once everyone's updated