@steviez 2021-07-14
Thanks for the suggestions to kickstart the brainstorm @ryoqun & @carllin. Agreed that we have to have two separate cf's to avoid mixing data-structures within one cf.

> > > This would be a breaking change since ShredIndex gets serialized/deserialized into/out of Rocks; I'm unsure how we should go about rolling it out.
> > 
> > 
> > You could introduce a separate column family to hold this new index structure, and then we can deprecate and delete the old index once everyone's updated. If we need to support the downgrade path, we'll have to run a converter though
> 
> yeah, downgrade option would be nice to have...:
> 
> so alternatively,
> 
> * write to the old & new cfs (bite the 2x temporal write?...)
> 
> or
> 
> * introduce feature flag switching which to use?

Is the idea with a feature flag something like this:
```
if new_index_feature_enabled {
    // Write to new cf
} else {
    // Write to old cf
}
```
If so, I think this could get messy.
- For the upgrade case, we'd need to have logic to fallback on reading the old cf if an item isn't in new column family
- For the downgrade case where validator ran for `d` slots, all of the other cf's would be up to `slot X`, but the old index cf would only be up to `slot X-d` which would cause issues

So, I think Ryo's suggestion of writing both (and absorbing the hit) might be necessary. To sum it up, I think it'd look like this:
1) Add a new cf (something like `index2` or `shred_index`)
2) Add a feature flag to write to both the existing and the new index cf's
3) Only read from the new index cf
   - This means old cf only populated to support possible rollback situation
4) At some point in the future (once all networks running whatever release this first goes into), strip out the old cf & feature flag (only write to new cf)
5) As part of previous step, may want to check for existence of old cf; if it exists, drop it in order to remove all remnants of old cf from blockstore
6) Remove the check from previous step once change from previous step propagated through all networks

So, 1-3 would be part of this PR, 4-5 would be to clean that up, and then step 6. to cleanup the cleanup haha? Does this sound right / reasonable?