@steviez 2021-07-20
@carllin @ryoqun - I updated the PR with logic to write both columns, not super pretty with some near-duplicate code. Some things:
- I named new column `Index2` with `ShredIndex2` sub-types; don't love it, but creative juices weren't flowing for anything better
- I commented out functions on original `Index` and `ShredIndex` objects; this made it easier to find places to update. I could uncomment these back out; I think leaving them commented out (or deleted) is the move to avoid someone unintentionally using the old CF
- One of the shortcomings of the new implementation was shown by unit test `test_should_insert_coding_shred()`. Part of this test attempts to insert a shred with `index = u32::MAX - 1`.
  - For a BTreeSet, this is fine as we just add another node in the tree. 
  - For a BitVec, this is trouble since the shred index is the same as the bit index in `BitVec`. In order to access `u32::MAX - 1`, that `BitVec` will need access to `2^32 bits == 2^29 bytes == 512 MB`. When I implemented this, I was using the below assumption as a bound. I can update the test but I'm wondering if we should more strictly enforce this in shred to avoid allowing the the `BitVec` from trying to claim all that memory
  https://github.com/solana-labs/solana/blob/b13caa14c95794aa44e2b7160731040accd87092/ledger/src/blockstore.rs#L87